fix(test): fix typing issues (#587)

+++ tests/functional/nccl_message_transfer.c
@@ -302,7 +302,7 @@ int main(int argc, char* argv[])
 						inflight_reqs--;
 						req_completed[idx] = 1;
 
-						if (received_size !=
+						if ((size_t)received_size !=
 						    NCCL_OFI_MIN(send_sizes[szidx], recv_sizes[szidx])) {
 							NCCL_OFI_WARN(
 								"Wrong received size %d (send size: %zu recv size %zu)",
